# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /studyplans:
    get:
      operationId: getStudyPlans
      tags: [studyplan]
      summary: List study plans
      responses:
        "200":
          description: Study plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyPlan"
    post:
      operationId: createStudyPlan
      tags: [studyplan]
      summary: Save a study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudyPlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
  /studyplans/{id}:
    delete:
      operationId: deleteStudyPlan
      tags: [studyplan]
      summary: Delete a study plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PROGRESS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StudyPlan:
      type: object
      properties:
        id:
          type: integer
        subjects:
          type: string
        examDate:
          type: string
        planData:
          type: string
        createdAt:
          type: string
      required: [id, subjects, examDate, planData, createdAt]
    CreateStudyPlanInput:
      type: object
      properties:
        subjects:
          type: string
        examDate:
          type: string
        planData:
          type: string
      required: [subjects, examDate, planData]

    # ─── PROGRESS ─────────────────────────────────────────────────────────────────
...